+++ CHANGELOG.md
@@ -1,3 +1,18 @@
+## 0.0.2
+
+* Add the example to the `README.md` file.
+* Add the `example` folder.
+* Fix the `CHANGELOG.md` file.
+* Update the `README.md` file.
+
 ## 0.0.1
 
-* TODO: Describe initial release.
+* Add the basic for the Realtime Database Utils class.
+* Get the reference of the root of the database.
+* Get the reference of the child in the database.
+* Add the data to the database.
+* Update the data in the database.
+* Remove the data from the database.
+* Get the data from the database.
+* Get the data from the database once.
+* Get the data from the database with a listener.